feat: introduce ParasutCoreModule for enhanced modularity
- Added ParasutCoreModule to encapsulate core functionalities and services.
- Updated ParasutModule to utilize ParasutCoreModule for better organization and maintainability.
- Refactored various modules to import ParasutCoreModule, ensuring consistent access to core services.
- Enhanced exports in the modules index to include ParasutCoreModule for improved accessibility.
