+++ lib/modules/contact/contact.module.ts
@@ -1,7 +1,9 @@
 import { Module } from "@nestjs/common";
+import { ParasutCoreModule } from "../../parasut-core.module";
 import { ParasutContactService } from "./contact.service";
 
 @Module({
+  imports: [ParasutCoreModule],
   providers: [ParasutContactService],
   exports: [ParasutContactService],
 })
