+++ lib/modules/expenses/bank-fee/bank-fee.module.ts
@@ -1,7 +1,9 @@
 import { Module } from "@nestjs/common";
+import { ParasutCoreModule } from "../../../parasut-core.module";
 import { ParasutBankFeeService } from "./bank-fee.service";
 
 @Module({
+  imports: [ParasutCoreModule],
   providers: [ParasutBankFeeService],
   exports: [ParasutBankFeeService],
 })
