+++ lib/modules/expenses/employee/employee.module.ts
@@ -1,7 +1,9 @@
 import { Module } from "@nestjs/common";
+import { ParasutCoreModule } from "../../../parasut-core.module";
 import { ParasutEmployeeService } from "./employee.service";
 
 @Module({
+  imports: [ParasutCoreModule],
   providers: [ParasutEmployeeService],
   exports: [ParasutEmployeeService],
 })
