+++ lib/modules/expenses/purchase-bill/purchase-bill.module.ts
@@ -1,7 +1,9 @@
 import { Module } from "@nestjs/common";
+import { ParasutCoreModule } from "../../../parasut-core.module";
 import { ParasutPurchaseBillService } from "./purchase-bill.service";
 
 @Module({
+  imports: [ParasutCoreModule],
   providers: [ParasutPurchaseBillService],
   exports: [ParasutPurchaseBillService],
 })
