+++ lib/modules/expenses/salary/salary.module.ts
@@ -1,7 +1,9 @@
 import { Module } from "@nestjs/common";
+import { ParasutCoreModule } from "../../../parasut-core.module";
 import { ParasutSalaryService } from "./salary.service";
 
 @Module({
+  imports: [ParasutCoreModule],
   providers: [ParasutSalaryService],
   exports: [ParasutSalaryService],
 })
