+++ lib/modules/expenses/tax/tax.module.ts
@@ -1,7 +1,9 @@
 import { Module } from "@nestjs/common";
+import { ParasutCoreModule } from "../../../parasut-core.module";
 import { ParasutTaxService } from "./tax.service";
 
 @Module({
+  imports: [ParasutCoreModule],
   providers: [ParasutTaxService],
   exports: [ParasutTaxService],
 })
