+++ lib/modules/formalization/formalization.module.ts
@@ -1,7 +1,9 @@
 import { Module } from "@nestjs/common";
+import { ParasutCoreModule } from "../../parasut-core.module";
 import { ParasutFormalizationService } from "./formalization.service";
 
 @Module({
+  imports: [ParasutCoreModule],
   providers: [ParasutFormalizationService],
   exports: [ParasutFormalizationService],
 })
