+++ lib/modules/sales/sales-invoice/sales-invoice.module.ts
@@ -1,7 +1,9 @@
 import { Module } from "@nestjs/common";
+import { ParasutCoreModule } from "../../../parasut-core.module";
 import { ParasutSalesInvoiceService } from "./sales-invoice.service";
 
 @Module({
+  imports: [ParasutCoreModule],
   providers: [ParasutSalesInvoiceService],
   exports: [ParasutSalesInvoiceService],
 })
