+++ lib/modules/sales/sales-offer/sales-offer.module.ts
@@ -1,7 +1,9 @@
 import { Module } from "@nestjs/common";
+import { ParasutCoreModule } from "../../../parasut-core.module";
 import { ParasutSalesOfferService } from "./sales-offer.service";
 
 @Module({
+  imports: [ParasutCoreModule],
   providers: [ParasutSalesOfferService],
   exports: [ParasutSalesOfferService],
 })
