+++ lib/modules/settings/category/category.module.ts
@@ -1,7 +1,9 @@
 import { Module } from "@nestjs/common";
+import { ParasutCoreModule } from "../../../parasut-core.module";
 import { ParasutCategoryService } from "./category.service";
 
 @Module({
+  imports: [ParasutCoreModule],
   providers: [ParasutCategoryService],
   exports: [ParasutCategoryService],
 })
