+++ lib/modules/settings/tag/tag.module.ts
@@ -1,7 +1,9 @@
 import { Module } from "@nestjs/common";
+import { ParasutCoreModule } from "../../../parasut-core.module";
 import { ParasutTagService } from "./tag.service";
 
 @Module({
+  imports: [ParasutCoreModule],
   providers: [ParasutTagService],
   exports: [ParasutTagService],
 })
